feat: add an option for separating multipleCheckbox question into a given number of columns
