+++ example/src/index.js
@@ -1,61 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
 import { ThemeProvider, merge } from 'theme-ui'
 import { ExampleTheme } from 'react-form-builder'
 
-const theme = {
-  buttons: {
-    primary: {
-      color: 'red'
-    },
-    customButton: {
-      color: 'green'
-    }
-  },
-  select: {
-    backgroundColor: 'red',
-    option: {
-      color: 'purple',
-      fontSize: '30px'
-    },
-    control: {
-      backgroundColor: 'gray',
-      borderRadius: '20px'
-    }
-  },
-  forms: {
-    container: {
-      contact: {
-        display: 'grid',
-        backgroundColor: 'green'
-      }
-    },
-    multipleImageCheckboxes: {
-      contact: {
-        display: 'grid',
-        gridTemplateColumns: '1fr 1fr',
-        label: {
-          div: {
-            backgroundColor: '#fff',
-            border: '1px solid #ccc',
-            borderRadius: '50%',
-            cursor: 'pointer',
-            height: '28px',
-            left: '0',
-            width: '28px',
-            svg: {
-              visibility: 'hidden'
-            }
-          }
-        }
-      }
-    }
-  },
-  label: {
-    color: 'brown'
-  }
-}
+import App from './App'
+import theme from './theme'
 
 ReactDOM.render(
   <ThemeProvider theme={merge(ExampleTheme, theme)}>
